requirements.txt
numpy
pandas
matplotlib
scikit-learn
joblib
tpot
pillow
protobuf
click
streamlit
seaborn